Creating Constants + Tools in uitils and updating the service

+++ src/main/java/ro/calin/Store/service/OrderService.java
@@ -13,12 +13,48 @@ public class OrderService {
     OrderRepository orderRepository;
 
     public boolean createOrder(Order order){
+        if(order.getOrderedProducts().size()>0 && order.getOrderedProducts().size()<=20 )
+        {
+            orderRepository.save(order);
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+
+    public boolean deleteOrder(Long idOrder){
+        if(orderRepository.findById(idOrder).isPresent())
+        {
+            orderRepository.deleteById(idOrder);
+            return true;
+        }
+        else
+            return false;
+    }
+
+    public boolean confirmOrder(Long idOrder){
+        Order order;
+        if(orderRepository.findById(idOrder).isPresent())
+             order = orderRepository.findById(idOrder).get();
+        else
+            return false;
+        order.setConfirmed(true);
         orderRepository.save(order);
         return true;
     }
 
-    public boolean deleteOrder(Long idOrder){
-        orderRepository.deleteById(idOrder);
+    public boolean orderDelivered(Long idOrder){
+        Order order;
+        if(orderRepository.findById(idOrder).isPresent())
+            order = orderRepository.findById(idOrder).get();
+        else
+            return false;
+        order.setDelivered(true);
+        orderRepository.save(order);
         return true;
     }
+
+
 }
